+++ common_pyutil/functional.py
@@ -6,28 +6,83 @@ def identity(x: Any):
     """Identity function.
 
     Example:
-        identity(10) == 10
-        identity(None) == None
+        >>> identity(10)
+        10
+        >>> identity(None)
+        None
     """
     return x
 
 
-def rpartial(func, *args):
+def rpartial(func: Callable, *args):
+    """Partial application of function `func` in reverse.
+
+    Args:
+        func: The function
+        args: Variable number of arguments
+
+    Example:
+        >>> def func(a, b, c,):
+        >>>     print(a, b, c)
+
+        # b and c take the place of 'b' and 'c' in the function arguments here
+        # That is, left to right
+        >>> rpartial(func, 2, 3)(1)
+        1, 2, 3
+
+    """
     def temp(*rest):
         return func(*rest, *args)
     return temp
 
 
-def maybe_is(kinds: List[type], x: Any):
+def maybe_is(x: Any, kinds: List[type]) -> Iterable[bool]:
+    """Return a map of checks for `x` in `kinds`.
+
+    Example:
+        >>> var = "var"
+        >>> maybe_is(var, [bool, str, type])
+        [False, True, False]
+    """
     return map(partial(apply, isinstance), [(x, k) for k in kinds])
 
 
-def maybe_(kinds: List[type], x: Any):
-    return last_item(first_by(zip(maybe_is(kinds, x), kinds), rpartial(nth, 0)))
+def maybe_(x: Any, kinds: List[type]) -> type:
+    """Return the type of `x` for the first value of which `isinstance` returns true
+    from a list of types in `kinds`.
+
+    Example:
+        >>> var = "var"
+        >>> maybe_(var, [bool, str, type])
+        str
+
+        >>> maybe_(str, [bool, str, type])
+        type
+    """
+    return last_item(first_by(zip(maybe_is(x, kinds), kinds), car))
 
 
-def maybe_then(kinds: List[type], x: Any, then: List[Callable]):
-    return last_item(first_by(zip(maybe_is(kinds, x), then), rpartial(nth, 0)))(x)
+def maybe_then(x: Any, kinds: List[type], then: List[Callable]) -> Any:
+    """Maybe if `x` is one of the `kinds`, call function `then` with that value.
+    Example:
+        >>> var = "var"
+        >>> kinds = [str, bool, type]
+        >>> funcs = [lambda x: len(x), lambda x: not x, lambda x: x.__qualname__]
+
+        >>> maybe_then(var, kinds, funcs)
+        3
+
+        >>> class Var:
+        >>>     pass
+        >>> var = Var()
+        >>> maybe_then(Var, kinds, funcs)
+        'Var'
+
+    Return:
+        The value from corresponding function or None if it's not one of the `kinds`.
+    """
+    maybe = first_by(zip(maybe_is(x, kinds), then), car)
+    return maybe and last_item(maybe)(x)
 
 
 def foldl(func: Callable, struct: Iterable):
@@ -35,7 +90,12 @@ def foldl(func: Callable, struct: Iterable):
     yield(func(next(it)))
 
 
-def first_by(struct: Optional[Iterable], by: Callable, predicate: Callable = identity):
+def first_by(struct: Optional[Iterable], by: Callable, predicate: Callable = identity) -> Any:
+    """Return first item in list on which `predicate` on output of `by` is True.
+
+    Return:
+        Either the found item or None
+    """
     if struct:
         it = iter(struct)
         while True:
@@ -47,7 +107,37 @@ def first_by(struct: Optional[Iterable], by: Callable, predicate: Callable = ide
                 return None
 
 
+def any_attr(obj: object, attrs: List[str],
+             predicate: Callable[[Any], bool] = identity):
+    """Return True if any attribute in `obj` satisfies `predicate` for a given
+    list of attributes `attrs`.
+
+    Args:
+        obj: Any `python` object
+        attrs: A `list` of names of attributes
+        predicat: A boolean function
+
+    """
+    return any(map(lambda x: predicate(getattr(obj, x, None)), attrs))
+
+
+def all_attrs(obj: object, attrs: List[str],
+              predicate: Callable[[Any], bool] = identity):
+    """Return True if all attributes in `obj` satisfy `predicate` for a given
+    list of attributes `attrs`.
+
+    Args:
+        obj: Any `python` object
+        attrs: A `list` of names of attributes
+        predicat: A boolean function
+
+    """
+    return all(map(lambda x: predicate(getattr(obj, x, None)), attrs))
+
+
 def last_item(struct: Optional[Iterable]):
+    """Return last item of `struct`.
+    """
     if struct:
         it = iter(struct)
         while True:
@@ -57,7 +147,7 @@ def last_item(struct: Optional[Iterable]):
                 return x
 
 
-def first(struct: Iterable, predicate: Callable = identity):
+def first(struct: Iterable, predicate: Callable):
     """Return first item of `struct` which satisfies `predicate`.
     """
     return first_by(struct, identity, predicate)
@@ -76,6 +166,8 @@ def first_item(struct: Iterable):
 
 
 def nth(struct: Iterable, indx: int):
+    """Return nth item of `struct`.
+    """
     it = iter(struct)
     i = 0
     x = None
@@ -89,16 +181,38 @@ def nth(struct: Iterable, indx: int):
 
 
 def applify(func: Callable, struct: Iterable[Iterable]):
+    """`` all the iters and apply function `func` to each of them
+    """
     return map(partial(apply, func), struct)
 
 
+def zip_with(func: Callable, *iters) -> Iterable:
+    """`zip` all the iters and apply function `func` to each of them
+    """
+    return map(func, zip(*iters))
+
+
 def apply_list(funcs: Iterable[Callable], struct: Iterable) -> List:
-    """Apply each function to item in two iterables.
+    """Apply each function in an Iterable to corresponding item in another.
+
+    Example:
+        >>> funcs = [lambda x: x ** 2, lambda x: 2 * x]
+        >>> its = [3, 4]
+        >>> apply_list(funcs, its)
+        [9, 8]
     """
     return [fn(it) for fn, it in zip(funcs, struct)]
 
 
 def apply(func: Callable, args: List):
+    """Call `func` expanding `args`.
+
+    Example:
+        >>> def add(a, b):
+        >>>     return a + b
+        >>> apply(add, [1, 2])
+        3
+    """
     return func(*args)
 
 
@@ -112,17 +226,19 @@ def pipe(*args: Callable) -> Callable:
     """Perform a function composition left to right.
 
     Example:
-        def f(x: int):
-            print(x)
-            return str(x)
+        >>> def f(x: int):
+        >>>     print(x)
+        >>>     return str(x)
+        >>>
+        >>> def g(x: str):
+        >>>     return "func g " + x
+        >>>
+        >>> def h(x: str):
+        >>>     return "func h " + x
 
-        def g(x: str):
-            return "func g " + x
-
-        def h(x: str):
-            return "func h " + x
-
-        pipe(f, g, h)(10)   # prints 10 and returns "func h func g 10"
+        >>> pipe(f, g, h)(10)
+        10
+        'func h func g 10'
 
     """
     return partial(reduce, lambda x, y: y(x), args)
@@ -196,7 +312,13 @@ def set_lens(obj: Optional[Dict], keys: List[str], val: Any) -> bool:
 
 
 def lens(obj: Optional[Dict], *args) -> Any:
-    """Return a value in a nested object.
+    """Recursively access values for given keys in :class:`dict` `obj`
+
+    Args:
+        obj: The object to scope
+        args: variable number of arguments
+
+    Returns None if a value isn't found for a sequence of keys.
     """
     if obj is None:
         return None
@@ -207,21 +329,21 @@ def lens(obj: Optional[Dict], *args) -> Any:
 
 
 def difference(a: Iterable, b: Iterable) -> set:
-    """Return set difference of two iterables"""
+    """Return :class:`set` difference of two iterables"""
     a = set([*a])
     b = set([*b])
     return a - b
 
 
 def intersection(a: Iterable, b: Iterable) -> set:
-    """Return intersection of two iterables"""
+    """Return :class:`set` intersection of two iterables"""
     a = set([*a])
     b = set([*b])
     return a.intersection(b)
 
 
 def union(a: Iterable, b: Iterable) -> set:
-    """Return union of two iterables"""
+    """Return :class:`set` union of two iterables"""
     a = set([*a])
     b = set([*b])
     return a.union(b)
@@ -235,7 +357,7 @@ def concat(list_var: Iterable[List]) -> List:
     return temp
 
 
-def takewhile(predicate, seq):
+def takewhile(predicate: Callable[[Any], bool], seq: Iterable) -> Iterable:
     """Lazily evaluated takewhile
 
     :param predicate: First failure of predicate stops the iteration. Should return bool
@@ -254,27 +376,28 @@ def takewhile(predicate, seq):
             return None
 
 
-def flatten_struct(struct, _type=None):
-    """Return a flattend list from a possibly nested iterable.
+# FIXME: Broken
+# def flatten_struct(struct: Iterable, _type=None):
+#     """Return a flattend iterable from a possibly nested iterable.
 
-    I'm not sure there's a usecase for this beyond lists
-    """
-    retval = []
-    # in case the underlying structure is also an iterable to avoid that also
-    # being extended with retval, e.g., a string
-    t = _type or type(struct)
-    it = iter(struct)
-    _next = it.__next__()
-    while _next:
-        try:
-            if not isinstance(_next, t):
-                retval.append(_next)
-            else:
-                retval.extend(flatten(_next, t))
-            _next = it.__next__()
-        except StopIteration:
-            return retval
-    return retval
+#     I'm not sure there's a usecase for this beyond lists
+#     """
+#     retval = []
+#     # in case the underlying structure is also an iterable to avoid that also
+#     # being extended with retval, e.g., a string
+#     t = _type or type(struct)
+#     it = iter(struct)
+#     _next = it.__next__()
+#     while _next:
+#         try:
+#             if not isinstance(_next, t):
+#                 retval.append(_next)
+#             else:
+#                 retval.extend(flatten(_next, t))
+#             _next = it.__next__()
+#         except StopIteration:
+#             return retval
+#     return retval
 
 
 def flatten(_list: List, depth: Optional[int] = None):
@@ -282,6 +405,22 @@ def flatten(_list: List, depth: Optional[int] = None):
 
     Args:
         depth: Depth to recurse
+
+    If depth is not given then the list is flattened as much as possible.
+
+    Example:
+        flatten([[0, 1], [2, 3]])
+        >>> [0, 1, 2, 3]
+
+        flatten([[0, 1], [2, [3, 4]]])
+        >>> [0, 1, 2, 3, 4]
+
+        flatten([[0, 1], [2, [3, 4]]], 1)
+        >>> [0, 1, 2, [3, 4]]
+
+        # Will not flatten a set
+        flatten([[0, 1], [2, {3, 4}]])
+        >>> [0, 1, 2, {3, 4}]
     """
     retval = []
     if depth is not None: