+++ ChangeLog.md
@@ -10,3 +10,10 @@
 
 ## [2021-05-18 Tue 05:54]
 - Added `apply_list` and `set_lens`.
+- Added types to `functional` module
+- Added `any_attr` and `all_attrs` in `functional` module.
+- Added `functional.zip_with`
+- Added `system.Semver`. Got annoyed with existing python semver library with
+  its warnings of breaking changes and all that. This is much simpler and I'm
+  sure will work well.
+- Version bump.